requirements.txt
streamlit
openai
pylint
matplotlib
google-generativeai   